Remove Noto Emoji font and change emojis
Despite using the Noto Emoji font as an alternative for displaying
the rock emoji, which appears to not be fully supported everywhere, some
of the emojis appear to not render in the style of the font in some
browsers, notably Firefox, which makes the game experience less ideal.

The hand-style emojis (and without the Noto Emoji font) seem to
show better support.